+++ TesteDesenvolvedor/TesteDesenvolvedor.Repository/Interface/IVeiculoRepository.cs
@@ -7,8 +7,8 @@ namespace TesteDesenvolvedor.Repository.Interface
 {
     public interface IVeiculoRepository : IRepository
     {
-        Task<Veiculo> FindByIdAsync(int id);
+        Task<Veiculo> FindByIdAsync(long id);
         Task<List<Veiculo>> GetAllAsync();
-        Task<List<Veiculo>> FindAllVeiculosByLinhasAsync(int linhaId);
+        Task<List<Veiculo>> FindAllVeiculosByLinhasAsync(long linhaId);
     }
 }
